Move api to pkg/requests
