+++ pkg/requests/requests_test.go
@@ -1,4 +1,4 @@
-package api
+package requests
 
 import (
 	"io/ioutil"
